+++ client/src/components/typing/short/PracticeShort.tsx
